Made some changes

+++ vars/creatingGcpInstance.groovy
@@ -1,4 +1,4 @@
-def call(Closure createGCPinstance) {
+def call(Map config=[:], Closure body) {
   pipeline {
     agent any
     stages {
@@ -15,32 +15,19 @@ def call(Closure createGCPinstance) {
               
         }
       }
-        stage('Destroy GCP instance') {
+      stage('Destroy GCP instance') {
             steps {
+            script {  
+            if (config.Destroy) {  
             timeout(time: 5, unit: 'MINUTES') {
             input(id: "Destroying terraform instance", message: "Are you sure to destroy ${params.project_name}?", ok: 'Destroy')
             echo 'Destroying GCP instance'
             sh 'terraform destroy -force'
+            }
+            }
             }
         }
       }
     }
   }
 }
-/*def call(Closure destroyGCPinstance) {
-  pipeline {
-    agent any
-    stages {
-      stage('Destroy GCP instance') {
-            steps {
-              timeout(time: 5, unit: 'MINUTES') {
-              input(id: "Destroying terraform instance", message: "Are you sure to destroy ${params.project_name}?", ok: 'Destroy')
-              
-              echo 'Destroying GCP instance'
-              sh 'terraform destroy -force'
-              }
-            }
-        }
-    }
-  }
-}*/
